fix: Renamed container and improved folder structure
