+++ app/src/components/market/sections/market_buy/scalar_market_buy.tsx
@@ -6,7 +6,7 @@ import ReactTooltip from 'react-tooltip'
 import styled from 'styled-components'
 
 import { STANDARD_DECIMALS } from '../../../../common/constants'
-import { SharedPropsInterface } from '../../../../pages/MarkeBuyContainerV2'
+import { SharedPropsInterface } from '../../../../pages/market_sections/MarkeBuyContainer'
 import { getLogger } from '../../../../util/logger'
 import { RemoteData } from '../../../../util/remote_data'
 import { calcXValue, formatBigNumber, formatNumber, getUnit } from '../../../../util/tools'
