+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components'
 import { useGraphMarketUserTxData } from '../../../../../hooks'
 import { WhenConnected, useConnectedWeb3Context } from '../../../../../hooks/connectedWeb3'
 import { useRealityLink } from '../../../../../hooks/useRealityLink'
-import { MarketBuyContainerV2 } from '../../../../../pages/MarkeBuyContainerV2'
+import { MarketBuyContainer } from '../../../../../pages/market_sections/MarkeBuyContainer'
 import { getNativeAsset, networkIds } from '../../../../../util/networks'
 import { getUnit, isDust } from '../../../../../util/tools'
 import { BalanceItem, MarketDetailsTab, MarketMakerData, OutcomeTableValue } from '../../../../../util/types'
@@ -387,7 +387,7 @@ const Wrapper = (props: Props) => {
         )}
         {currentTab === MarketDetailsTab.history && <MarketHistoryContainer marketMakerData={marketMakerData} />}
         {currentTab === MarketDetailsTab.buy && (
-          <MarketBuyContainerV2
+          <MarketBuyContainer
             fetchGraphMarketMakerData={fetchGraphMarketMakerData}
             fetchGraphMarketUserTxData={fetchGraphMarketUserTxData}
             isScalar={isScalar}
